refactor: relocate NoopSearchClientsProxy to search-client module
It is useful to have a default noop implementation to use by dependants like
zeebe-process-test, instead of having to adapt a local duplicated NoopClient to changes.
